25-03-2024

+++ src/main/java/com/zubayer/security/SpringSecurityAuditorAware.java
@@ -14,11 +14,11 @@ import com.zubayer.model.MyUserDetail;
  * @since Mar 24, 2024
  * CSE202101068
  */
-public class SpringSecurityAuditorAware implements AuditorAware<Long> {
+public class SpringSecurityAuditorAware implements AuditorAware<Integer> {
 
 	@Override
-	public Optional<Long> getCurrentAuditor() {
-		Long id = Long.valueOf(0);
+	public Optional<Integer> getCurrentAuditor() {
+		Integer id = Integer.valueOf(0);
 		MyUserDetail user = getLoggedInUserDetails();
 		if(user != null && StringUtils.isNotBlank(user.getUsername())) id = user.getId();
 		return Optional.ofNullable(id);
